fix: price typing on advertisement and all of its use cases + add FIPEClientInterface

+++ redline-core/src/main/java/com/yuriolivs/redlinecore/domain/advertisement/Advertisement.java
@@ -134,7 +134,7 @@ public class Advertisement {
         return scoreHistory.get(scoreHistory.size() - 1).getValue();
     }
 
-    public BigDecimal getPrice() {
+    public Double getPrice() {
         return priceHistory.get(priceHistory.size() - 1).getPrice();
     }
